Add base class for frame buffer
